+++ Assets/WorldGeneration/_Scripts/ScriptableObjects/NoiseSettings.cs
@@ -6,8 +6,8 @@ namespace WorldGeneration._Scripts.ScriptableObjects
     public class NoiseSettings : ScriptableObject
     {
         // Seed
-        [Header("Seed")] public bool useRandomSeed;
-        public string seed = "Hello World!";
+        // [Header("Seed")] public bool useRandomSeed;
+        // public string seed = "Hello World!";
         // [Range(10000.0f, 1000000.0f)] public float seedScale = 100000.0f;
 
         // Gradient noise settings