+++ Assets/WorldGeneration/_Scripts/TerrainGeneration/WaterGenerator.cs
@@ -20,12 +20,13 @@ namespace WorldGeneration._Scripts.TerrainGeneration
             if (Instance == null)
             {
                 transform.parent = null;
-                DontDestroyOnLoad(gameObject);
+                // DontDestroyOnLoad(gameObject);
                 Instance = this;
             }
             else if (Instance != this)
             {
-                Destroy(gameObject);
+                Destroy(Instance);
+                Instance = this;
             }
         }